Project 2/requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
bcrypt==4.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
config==0.5.1
contourpy==1.3.1
CTkMessagebox==2.7
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
fonttools==4.54.1
google-ai-generativelanguage==0.6.10
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
grpcio==1.67.1
grpcio-status==1.67.1
httplib2==0.22.0
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
numpy==2.1.3
oauthlib==3.2.2
packaging==24.1
pefile==2023.2.7
pillow==11.0.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
PyJWT==2.9.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pywin32-ctypes==0.2.3
reportlab==4.2.5
requests==2.25.1
requests-oauthlib==2.0.0
rsa==4.9
setuptools==75.3.0
six==1.16.0
tqdm==4.66.6
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.6
Werkzeug==3.1.3
